sync input value and $viewValue, use ngModel.$render
This fixes a bug that was revealed during enhancements to the timeDatePicker meta directive.
Lesson: If the $viewValue of ngModel gets out of sync with the input value of the element some strange things can happen.
Explanation: The previous version of the blur event handler directly updated the input value without updating the $viewValue.
After a programmatic update to the model, the new $viewValue was different from the input value but the same as the out of sync $viewValue.
This prevented the the input value from being updated because there was no delta in the $viewValue.
see also: sport_admin/pull/1894 commit df8da01